Fully working fully tested with no ring buffer (8MHz)
This version has been fully tested and cleaned. So should be close to
final version. I have removed the ring buffer because it is not
needed with a bus at 4/8 MHz. I also remove the UARTBase class
and moved the read/write code to the channel class...

Changes to be committed:
	modified:   esphome/components/sc16is75x_spi/sc16is75x_spi.cpp
	modified:   esphome/components/sc16is75x_spi/sc16is75x_spi.h